Move the setting-path predicate beneath both of its declaration readers
The live accessor's lazy credentials import gave capabilities and
credentials a mutual dependency (pylint R0401). is_renderable_setting_path
and its two constants move to models.py — beneath both the capability and
the store-credential declarations that check paths against it — so the
dependency runs one way again; capabilities re-exports the name for its
existing consumers. Also the format/lint round: module-qualified private
access in tests, a typed opt-in helper in the service tests, and a scoped
pylint disable on the accessor's deliberate lazy client imports.

Seat: lead-developer
